feat: add Autoterm emulator for testing without hardware
New library: lib/AutotermEmulator/
  - Simulates Autoterm Flow 5D heater ECU on Serial1
  - Full state machine: Off → Starting → Warming → Running →
    ShuttingDown → Cooling → Off (with realistic timing)
  - Thermal simulation: core temp ramps up/down based on state
    and power level (0-9)
  - Battery voltage simulation with running dip and jitter
  - Error injection for testing safety layer (overheat, voltage,
    glow plug, flame, ignition, fan failures)

Internal GPIO loopback — NO external wiring needed:
  - Driver on Serial2 (RX=16, TX=17)
  - Emulator on Serial1 (RX=17, TX=16)
  - ESP32 GPIO matrix routes TX→RX internally on the same pins

New PlatformIO environment [env:emulator]:
  - Build: pio run -e emulator
  - Flash: pio run -e emulator -t upload
  - Defines -D EMULATOR_MODE, excludes OneWire/DallasTemperature

Emulator serial commands (in addition to driver commands):
  coolant <°C>     — set simulated coolant temperature
  error <type>     — inject heater error (overheat/voltage/flame/...)
  error clear      — clear injected error
  voltage <V>      — override battery voltage
  voltage normal   — resume voltage simulation
  emu status       — show emulator internals

Also fixed:
  - AUTOTERM_DATA_BITS type: uint8_t → uint32_t (SERIAL_8N1 overflow)
  - Guarded heater_addon.cpp with #ifndef EMULATOR_MODE

Build: both esp32dev and emulator environments compile clean

+++ lib/AutotermUART/autoterm_protocol.h
@@ -19,7 +19,7 @@
 // Physical layer
 // ---------------------------------------------------------------------------
 static constexpr uint32_t AUTOTERM_BAUD        = 2400;
-static constexpr uint8_t  AUTOTERM_DATA_BITS   = SERIAL_8N1;
+static constexpr uint32_t AUTOTERM_DATA_BITS   = SERIAL_8N1;
 
 // ---------------------------------------------------------------------------
 // Frame constants
